Part C attempt: This version reads one line of code from the user.
