--- PartC/lexemes_data.txt
@@ -1,11 +1,0 @@
-".*?" STRING
-begin|for|each|in|endfor|print|end KEYWORD
-\b[a-zA-Z_]\w* ID
-\b[\d]+\b NUM
-[+\-*/] ARITHMETIC_OP
-= ASSIGN_OP
-, SEPARATOR
-\[ L_BRACKET
-\] R_BRACKET
-: COLON_OP
-; DELIMITER
