+++ Umaya_PartC/MainWindow.py
@@ -1,0 +1,9 @@
+from PySide6.QtWidgets import (
+    QMainWindow, QApplication, QVBoxLayout,
+    QPushButton, QWidget, QHBoxLayout,
+)
+
+class MainWindow(QMainWindow):
+    def __init__(self, /):
+        super().__init__()
+
